pyproject.toml
[tool.poetry]
name = "siros"
version = "0.1.0"
description = "SIROS stuff"
authors = ["Milton Abrunhosa <miltonmla@decea.mil.br>"]
license = "CC0 1.0 Universal"

[tool.poetry.dependencies]
python = "^3.7"
websocket-client = "^1.3.2"
stomper = "^0.4.3"
python-dotenv = "^0.20.0"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
mypy = "^0.950"
pylint = "^2.13.8"
autopep8 = "^1.6.0"
types-requests = "^2.27.25"
flake8 = "^4.0.1"
pytest-coverage = "^0.0"
pip-audit = "^2.2.1"
safety = "^1.10.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
